Moving index field type definitions to the Index definition from Models
This is important, because DynamoDB does not allow mixed data types for
index fields (attributes). In other words, all models must follow the
schema definition when it comes to index fields.

Because of this, having the type definition on the model itself just
becomes extra boilerplate AND has the potential to lead to incorrect
behaviour, when a user does not know this limitation and tries to define
different types on different models but for the same index.