Adding update support
All update operations should work fine, there might be some edge-cases
with data validation at the correct layers, but generally speaking, it
is working as intended.

Two interfaces again, one directly on the table where the user needs to
provide the class to deserialize data into and one for convinience on the
model class itself.