Adding support for index field ordering
Relying on the order in which these properties are defined on a model
is fairly error-prone and requires deep knowledge of this library.

Exposing a specific order field for index fields makes it much clearer
how the sort key value is going to be constructed and makes the usage
of this library safer.